Place ClusterEvents and refactoring
