+++ src/components/fetching/ClusterEvents.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import EventListFetch, { EventFetchProps } from '../ui/EventListFetch';
+import EventListFetch, { EventFetchProps } from './EventListFetch';
 
 const ClusterEvents: React.FC<EventFetchProps> = ({ entityId }) => (
   <EventListFetch entityId={entityId} entityKind="cluster" />
